revert headerdata changes
